fix: rebase onto main + address CodeRabbit review feedback
- Resolve merge conflict with pagination changes (fetchAllPaginated)
- Fix type mismatch: session_id column integer→bigint to match sessions.id
- Fix dashboard fallback: don't fall back to branch heuristic when sessionId
  is set but not found (prevents attaching wrong session metadata)
- Fix CLI: main() now uses syncSessionFile() to include FK linking
- Add clobber guard: skip FK update if agent session already has sessionId
- Add FK constraint error handling in PATCH endpoint (400 vs 500)

+++ apps/wiki-server/drizzle/0063_add_session_id_to_agent_sessions.sql
@@ -15,6 +15,6 @@
 -- pattern impossible. agent_sessions is a small table (hundreds of rows), so the
 -- brief SHARE ROW EXCLUSIVE lock is acceptable.
 
-ALTER TABLE "agent_sessions" ADD COLUMN IF NOT EXISTS "session_id" integer REFERENCES "sessions"("id") ON DELETE SET NULL;
+ALTER TABLE "agent_sessions" ADD COLUMN IF NOT EXISTS "session_id" bigint REFERENCES "sessions"("id") ON DELETE SET NULL;
 
 CREATE INDEX IF NOT EXISTS "idx_as_session_id" ON "agent_sessions" ("session_id");
